--- src/tests/text/nameOfFinderTests.ts
@@ -1,62 +1,0 @@
-﻿import * as assert from "assert";
-import { NameOfFinder, StringIterator } from "../../text";
-import { testReplaceInfo } from "./helpers";
-
-describe("NameOfFinder", () => {
-    describe("#indexOfAll", () => {
-        describe("at beginning and end of string", () => {
-            const finder = new NameOfFinder(new StringIterator("nameof(str);nameof(t)"));
-            const indexes = finder.indexOfAll();
-
-            describe("nameof: 0", () => {
-                testReplaceInfo(indexes[0], {
-                    typeArgText: "",
-                    args: ["str"],
-                    pos: 0,
-                    end: 11,
-                    showFull: false
-                });
-            });
-            describe("nameof: 1", () => {
-                testReplaceInfo(indexes[1], {
-                    typeArgText: "",
-                    args: ["t"],
-                    pos: 12,
-                    end: 21,
-                    showFull: false
-                });
-            });
-        });
-
-        describe("name tests", () => {
-            function startStringTest(startString: string) {
-                const finder = new NameOfFinder(new StringIterator(startString + "nameof(str);"));
-                const indexes = finder.indexOfAll();
-
-                it("should not find anything", () => {
-                    assert.equal(indexes.length, 0);
-                });
-            }
-
-            describe("where there's an alpha char at the beginning", () => {
-                startStringTest("a");
-            });
-
-            describe("where there's a number char at the beginning", () => {
-                startStringTest("0");
-            });
-
-            describe("where there's an underscore char at the beginning", () => {
-                startStringTest("_");
-            });
-
-            describe("where it's a property of an object", () => {
-                startStringTest("test.");
-            });
-
-            describe("where it's a property of an object, but there's a space after the period", () => {
-                startStringTest("test. \t\r");
-            });
-        });
-    });
-});
