--- src/text/ReplaceInfo.ts
@@ -1,7 +1,0 @@
-﻿export interface ReplaceInfo {
-    pos: number;
-    end: number;
-    readonly args: string[];
-    readonly typeArgText: string;
-    showFull: boolean;
-}
