+++ src/text/index.ts
@@ -1,10 +1,4 @@
 export * from "./getFileNamesFromGlobs";
-export * from "./handleComment";
-export * from "./handleNameOf";
-export * from "./NameOfFinder";
-export * from "./replaceCallExpressionReplacesInText";
 export * from "./replaceInFiles";
-export * from "./ReplaceInfo";
 export * from "./replaceInText";
 export * from "./stream";
-export * from "./StringIterator";
