+++ src/transformation/common.ts
@@ -1,0 +1,3 @@
+export function throwError(message: string): never {
+    throw new Error(`[ts-nameof]: ${message}`);
+}
