+++ src/transformation/index.ts
@@ -1,1 +1,4 @@
+export * from "./checks";
+export * from "./common";
 export * from "./transformerFactory";
+export * from "./transforms";
